backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pymysql>=1.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
requests>=2.28.0
jinja2>=3.1.0
